Move _get_output_filename() into the base class
This will facilitate reuse of the same method in the EmailViolations
class.
